disable general Facebook test
The tests fail with the following error if run with PHP >=7.4:

  Trying to access array offset on value of type null

The error is caused by Backend/Facebook.php L34:

  $accessToken = urlencode($this->config['app_id'])
    .'|'.urlencode($this->config['secret']);

$this->config is NULL in ShariffTest::testShariff and
ShariffTest::testCacheOptions because the 'Facebook' key does not
exist in the supplied test configurations.

I _think_ before PHP 7.4. this happened:
Backend/Facebook.php L34 would not throw an error but result in an invalid
accessToken, which would result in an invalid Facebook Graph API URL. Calling
this URL would result in an invalid API response which would result in an
empty share count for Facebook. The tests 'worked', because the Facebook service
was loaded but its return values were never checked.

A problem is that Backend/Facebook.php tries to fail in its setConfig
function if the API credentials are missing. But the Backend/ServiceFactory.php
only calls setConfig if the service has actually been configured. So not
configuring the Facebook service but loading it results in an unconfigured
service that cannot be used and generates invalid accessTokens and
Graph API URLs.